Fix lobbies.js require references, move jsonUtil to util folder
